Removed main buffer rendering. Should be faster.
In the past, the main buffer had to be rendered after every update. Now,
the main buffer is written to but rarely rendered because it is also a
wrapper for the base terminal object.

This allows the program to skip a level of additional drawing, making
the program feel smoother and more responsive.

It also allows for some features that were previously difficult to
implement due to the aforementioned re-rendering of the main buffer.
Things like rednet message/email notifications, for example.